fix(runtime): address computer manager review
